backend-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.3.7
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
aiofiles==23.2.1
Pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2

